Rename HttpService and interface
