+++ lib/src/main/kotlin/com/swmansion/starknet/provider/gateway/GatewayProvider.kt
@@ -10,10 +10,10 @@ import com.swmansion.starknet.extensions.put
 import com.swmansion.starknet.extensions.toDecimal
 import com.swmansion.starknet.provider.Provider
 import com.swmansion.starknet.provider.Request
-import com.swmansion.starknet.service.RequestProcessingService
-import com.swmansion.starknet.service.RequestProcessingService.Payload
-import com.swmansion.starknet.service.http.HttpRequest
 import com.swmansion.starknet.service.http.HttpService
+import com.swmansion.starknet.service.http.HttpService.Payload
+import com.swmansion.starknet.service.http.HttpRequest
+import com.swmansion.starknet.service.http.OkhttpHttpService
 import com.swmansion.starknet.service.http.handlers.BasicHttpErrorHandler
 import kotlinx.serialization.json.*
 
@@ -29,7 +29,7 @@ class GatewayProvider(
     private val gatewayUrl: String,
     override val chainId: StarknetChainId,
 ) : Provider {
-    private val httpService: RequestProcessingService = HttpService(BasicHttpErrorHandler())
+    private val httpService: HttpService = OkhttpHttpService(BasicHttpErrorHandler())
 
     @Suppress("SameParameterValue")
     private fun gatewayRequestUrl(method: String): String {
