+++ lib/src/main/kotlin/com/swmansion/starknet/provider/rpc/JsonRpcProvider.kt
@@ -8,9 +8,9 @@ import com.swmansion.starknet.extensions.add
 import com.swmansion.starknet.extensions.put
 import com.swmansion.starknet.provider.Provider
 import com.swmansion.starknet.provider.Request
-import com.swmansion.starknet.service.RequestProcessingService
-import com.swmansion.starknet.service.http.HttpRequest
 import com.swmansion.starknet.service.http.HttpService
+import com.swmansion.starknet.service.http.HttpRequest
+import com.swmansion.starknet.service.http.OkhttpHttpService
 import com.swmansion.starknet.service.http.handlers.BasicHttpErrorHandler
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.json.*
@@ -25,7 +25,7 @@ class JsonRpcProvider(
     private val url: String,
     override val chainId: StarknetChainId,
 ) : Provider {
-    private val httpService: RequestProcessingService = HttpService(BasicHttpErrorHandler())
+    private val httpService: HttpService = OkhttpHttpService(BasicHttpErrorHandler())
 
     private fun buildRequestJson(method: String, paramsJson: JsonElement): Map<String, JsonElement> {
         val map = mapOf(
@@ -46,7 +46,7 @@ class JsonRpcProvider(
         val requestJson = buildRequestJson(method.methodName, paramsJson)
 
         val payload =
-            RequestProcessingService.Payload(url, "POST", emptyList(), requestJson.toString())
+            HttpService.Payload(url, "POST", emptyList(), requestJson.toString())
 
         val jsonRpcDeserializer = JsonRpcResponseDeserializer(responseSerializer)
