+++ lib/src/main/kotlin/com/swmansion/starknet/service/http/HttpRequest.kt
@@ -1,16 +1,15 @@
 package com.swmansion.starknet.service.http
 
 import com.swmansion.starknet.provider.Request
-import com.swmansion.starknet.service.RequestProcessingService
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.json.Json
 import java.util.concurrent.CompletableFuture
 
 internal class HttpRequest<T>(
-    private val payload: RequestProcessingService.Payload,
+    private val payload: HttpService.Payload,
     // TODO: Probably it could be abstracted, to not depend on kotlinx serialization
     private val deserializer: DeserializationStrategy<T>,
-    private val service: RequestProcessingService,
+    private val service: HttpService,
 ) : Request<T> {
     private val json = Json {
         ignoreUnknownKeys = true
